requirements.txt
Flask==3.0.2
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
python-dotenv==1.0.1
Werkzeug==3.0.1 